requirements.txt
streamlit
openai
pandas
numpy
Pillow
tesserocr
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
